[Rename] : DTO 클래스 로케이션 수정에 따른 import 수정
