+++ src/main/java/com/sctk/cmc/common/dto/designer/DesignerJoinParam.java
@@ -1,10 +1,12 @@
-package com.sctk.cmc.service.dto.designer;
+package com.sctk.cmc.common.dto.designer;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
-@AllArgsConstructor
 @Getter
+@NoArgsConstructor
+@AllArgsConstructor
 public class DesignerJoinParam {
     private String name;
     private String nickname;
