+++ src/main/java/com/sctk/cmc/common/dto/designer/DesignerLoginParam.java
@@ -1,9 +1,13 @@
-package com.sctk.cmc.dto.designer;
+package com.sctk.cmc.common.dto.designer;
 
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
+@NoArgsConstructor
+@AllArgsConstructor
 @Builder
 public class DesignerLoginParam {
     private String email;
