+++ src/main/java/com/sctk/cmc/common/dto/member/MemberLoginParam.java
@@ -1,9 +1,13 @@
-package com.sctk.cmc.dto.member;
+package com.sctk.cmc.common.dto.member;
 
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
+@NoArgsConstructor
+@AllArgsConstructor
 @Builder
 public class MemberLoginParam {
     private String email;
