+++ src/main/java/com/sctk/cmc/web/dto/designer/DesignerInfoResponse.java
@@ -1,19 +1,21 @@
-package com.sctk.cmc.web.dto;
+package com.sctk.cmc.web.dto.designer;
 
-import com.sctk.cmc.service.dto.member.BodyInfoView;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @EqualsAndHashCode
-public class MemberInfoResponse {
+public class DesignerInfoResponse {
     private String name;
     private String profileImgUrl;
-    private BodyInfoView bodyInfoView;
+    private String introduce;
+    private int likes;
+    private List<String> highCategoryNames;
+    private List<String> lowCategoryNames;
 }
-
-
