+++ src/test/java/com/sctk/cmc/web/controller/MemberControllerTest.java
@@ -9,9 +9,9 @@ import com.sctk.cmc.service.abstractions.MemberService;
 import com.sctk.cmc.service.dto.member.BodyInfoView;
 import com.sctk.cmc.service.dto.member.MemberDetail;
 import com.sctk.cmc.service.dto.member.MemberInfo;
-import com.sctk.cmc.web.dto.BodyInfoPostRequest;
-import com.sctk.cmc.web.dto.MemberDetailResponse;
-import com.sctk.cmc.web.dto.MemberInfoResponse;
+import com.sctk.cmc.web.dto.member.BodyInfoPostRequest;
+import com.sctk.cmc.web.dto.member.MemberDetailResponse;
+import com.sctk.cmc.web.dto.member.MemberInfoResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
